requirements.txt
openai==1.35.10
spotipy==2.23.0
python-dotenv==1.0.1
